Add TUV-x version to the Julia interface; modularize the bindings
Start the progressive TUV-x scaffolding in Julia with the version string,
and restructure the bindings so components are added without preprocessor
guards.

- Split julia/bindings/musica_julia.cpp into per-component translation units
  (micm.cpp, tuvx.cpp) that self-register via a small registry
  (registration.hpp). The module entry point binds the MUSICA version and runs
  every registered component; no #ifdefs anywhere.
- tuvx.cpp is compiled only when MUSICA_ENABLE_TUVX is on (CMake target_sources,
  mirroring the Python bindings), and binds get_tuvx_version() ->
  musica::TUVX::GetVersion().
- Enable TUV-x in the Julia CI build (it is a Fortran library): install
  NetCDF + Fortran on Linux/macOS and flip MUSICA_ENABLE_TUVX=ON.
- Export get_tuvx_version and add a version test, mirroring the MUSICA version.

Per project decision, the Julia TUV-x tests only ever pass valid inputs, since
TUV-x calls Fortran STOP on bad config (which hangs on Windows runners).
