+++ julia/test/runtests.jl
@@ -11,6 +11,14 @@ using Musica
         println("MUSICA version: ", version)
     end
 
+    @testset "TUV-x version" begin
+        tuvx_version = Musica.get_tuvx_version()
+        @test tuvx_version isa AbstractString
+        @test !isempty(tuvx_version)
+        @test occursin(r"^\d+\.\d+\.\d+", tuvx_version)
+        println("TUV-x version: ", tuvx_version)
+    end
+
     @testset "CUDA availability" begin
         @test Musica.cpp_is_cuda_available() isa Bool
     end
